Created Service Classes, updated BookController accordingly, then also added more Mapping for increased functionality. Also updated the User model to allow percentage read of books and read status of books to be tracked for each user independently by creating a UserBooks object.

+++ src/main/java/com/bookreviews/Repositories/ReviewRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ReviewRepository extends CrudRepository<Review, Integer> {
+public interface ReviewRepository extends CrudRepository<Review, Long> {
 
     public List<Review> findByDisplayName(String displayName);
